Polymorphism on alarms, updated UI.
Distinguishing between repeating and non-repeating alarms to make
everything a hell of a lot easier.
