+++ Team4Clock/MainWindow.xaml
@@ -18,7 +18,8 @@
         <Label x:Name="TImeLabel" Content="Label" FontSize="150" VerticalAlignment="Center" Canvas.Left="201" Canvas.Top="157" Foreground="Black" Margin="0,154,0,170" HorizontalAlignment="Center" />
         <Button x:Name="snoozeButton" Content="Snooze" Visibility="Hidden" HorizontalAlignment="Center" Margin="272,369,409,121" VerticalAlignment="Center" Width="87"  Height="44" Click="snooze_Click" x:FieldModifier ="public" RenderTransformOrigin="-0.719,0.732" />
         <Button x:Name="awakeButton" Content="Wake Up" Visibility="Hidden" HorizontalAlignment="Center" Margin="381,369,300,121" VerticalAlignment="Center" Width="87" Height="44" Click="awake_Click" />
-        <Button x:Name="setAlarmBtn" Content="Set Alarm" HorizontalAlignment="Right" Margin="0,10,10,0" VerticalAlignment="Top" Width="87" Click="setAlarmBtn_Click" Height="44"/>
+        <Button x:Name="setAlarmBtn" Content="New Alarm" HorizontalAlignment="Right" Margin="0,10,10,0" VerticalAlignment="Top" Width="141" Click="setAlarmBtn_Click" Height="44"/>
+        <Button x:Name="rptAlarmBtn" Content="New Repeating Alarm" HorizontalAlignment="Right" Margin="0,72,10,0" VerticalAlignment="Top" Width="141" Click="rptAlarmBtn_Click" Height="44"/>
         <ToggleButton IsChecked="{Binding IsOpen, ElementName=alarmListFlyout}" Cursor="Hand" x:Name="List_Button" Content="List" Margin="10,10,0,0" VerticalAlignment="Top" Click="List_Click" Width="87" Height="44" HorizontalAlignment="Left"/>
         <Controls:Flyout x:Name="alarmListFlyout" Theme="Accent">
             <Controls:Flyout.Resources>
