+++ Team4Clock/SetAlarm.xaml.cs
@@ -22,115 +22,10 @@ namespace Team4Clock
         private bool isPm = false;        // default to AM
         private MainWindow mw = new MainWindow(); // The parent view object
 
-        // Mappings between buttons and days of week to simplify event handling
-        private Dictionary<CheckBox, DayOfWeek> buttonToDay;
-        private Dictionary<DayOfWeek, CheckBox> dayToButton;
-
-        // Collections of strings for use in ComboBoxes for alarm repeats. Set by initComboBoxes() function.
-        private ObservableCollection<string> hrsList = new ObservableCollection<string>();
-        private ObservableCollection<string> minsList = new ObservableCollection<string>();
-        private ObservableCollection<string> amPmList = new ObservableCollection<string>();
-
-
         public SetAlarm(MainWindow newMW)
         {
             this.mw = newMW; // The parent view is set 
             InitializeComponent();
-            initDictionaries();
-
-            initComboBoxes();
-        }
-
-        /* Initializes dictionaries used by this class.
-         * 
-         * This provides fast lookup between days of the week and their corresponding buttons.
-         */
-        private void initDictionaries()
-        {
-            // Button->Day mappings
-            buttonToDay = new Dictionary<CheckBox, DayOfWeek> {
-                {sunBtn, DayOfWeek.Sunday},
-                {monBtn, DayOfWeek.Monday},
-                {tueBtn, DayOfWeek.Tuesday},
-                {wedBtn, DayOfWeek.Wednesday},
-                {thuBtn, DayOfWeek.Thursday},
-                {friBtn, DayOfWeek.Friday},
-                {satBtn, DayOfWeek.Saturday}
-            };
-
-            // Day->Button mappings
-            dayToButton = new Dictionary<DayOfWeek, CheckBox>();
-            foreach (var entry in buttonToDay)
-            {
-                dayToButton.Add(entry.Value, entry.Key);
-            }
-        }
-
-        private void initComboBoxes()
-        {
-            // Set up shared string collections for combo boxes
-            for (int i = 1; i <= 12; i++)
-            {
-                hrsList.Add(i.ToString());
-            }
-            for (int i = 0; i < 60; i++)
-            {
-                minsList.Add(i.ToString("D2"));
-            }
-            amPmList.Add("AM");
-            amPmList.Add("PM");
-
-            // Set combo boxes to use shared sources (there's gotta be a better way to do this...)
-            this.sunHrs.ItemsSource = hrsList;
-            this.sunMins.ItemsSource = minsList;
-            this.sunAmPm.ItemsSource = amPmList;
-            this.monHrs.ItemsSource = hrsList;
-            this.monMins.ItemsSource = minsList;
-            this.monAmPm.ItemsSource = amPmList;
-            this.tueHrs.ItemsSource = hrsList;
-            this.tueMins.ItemsSource = minsList;
-            this.tueAmPm.ItemsSource = amPmList;
-            this.wedHrs.ItemsSource = hrsList;
-            this.wedMins.ItemsSource = minsList;
-            this.wedAmPm.ItemsSource = amPmList;
-            this.thuHrs.ItemsSource = hrsList;
-            this.thuMins.ItemsSource = minsList;
-            this.thuAmPm.ItemsSource = amPmList;
-            this.friHrs.ItemsSource = hrsList;
-            this.friMins.ItemsSource = minsList;
-            this.friAmPm.ItemsSource = amPmList;
-            this.satHrs.ItemsSource = hrsList;
-            this.satMins.ItemsSource = minsList;
-            this.satAmPm.ItemsSource = amPmList;
-
-            // Set up defaults
-            sunHrs.SelectedItem = "12";
-            sunMins.SelectedIndex = 0;
-            sunAmPm.SelectedIndex = 0;
-
-            monHrs.SelectedItem = "12";
-            monMins.SelectedIndex = 0;
-            monAmPm.SelectedIndex = 0;
-
-            tueHrs.SelectedItem = "12";
-            tueMins.SelectedIndex = 0;
-            tueAmPm.SelectedIndex = 0;
-
-            wedHrs.SelectedItem = "12";
-            wedMins.SelectedIndex = 0;
-            wedAmPm.SelectedIndex = 0;
-
-            thuHrs.SelectedItem = "12";
-            thuMins.SelectedIndex = 0;
-            thuAmPm.SelectedIndex = 0;
-
-            friHrs.SelectedItem = "12";
-            friMins.SelectedIndex = 0;
-            friAmPm.SelectedIndex = 0;
-
-            satHrs.SelectedItem = "12";
-            satMins.SelectedIndex = 0;
-            satAmPm.SelectedIndex = 0;
         }
 
         // Removes the current set alarm view from the stack allowing
@@ -246,10 +141,7 @@ namespace Team4Clock
             hours = (isPm) ? (hours + 12) : hours;
             hours = (hours == 24) ? 0 : hours;
             TimeSpan alarmTime = new TimeSpan(hours, Convert.ToInt32(min), 0);
-            Alarm alarm = new Alarm(alarmTime);
-
-            // Get repeat days and update the alarm with these days
-            setAlarmRepeats(alarm);
+            BasicAlarm alarm = new BasicAlarm(alarmTime);
 
             mw.setList(alarm);
         }
@@ -268,45 +160,6 @@ namespace Team4Clock
             hours += setPm ? 12 : 0;
             int mins = Int32.Parse(minsBox.SelectedItem.ToString());
             return new TimeSpan(hours, mins, 0);
-        }
-
-        private void setAlarmRepeats(Alarm alarm)
-        {
-            TimeSpan time;
-            if ((bool) sunBtn.IsChecked) {
-                time = parseRepeatBoxes(sunHrs, sunMins, sunAmPm);
-                alarm.SetRepeat(DayOfWeek.Sunday, /*repeats*/ true, time);
-            }
-            if ((bool)monBtn.IsChecked)
-            {
-                time = parseRepeatBoxes(monHrs, monMins, monAmPm);
-                alarm.SetRepeat(DayOfWeek.Monday, /*repeats*/ true, time);
-            }
-            if ((bool)tueBtn.IsChecked)
-            {
-                time = parseRepeatBoxes(tueHrs, tueMins, tueAmPm);
-                alarm.SetRepeat(DayOfWeek.Tuesday, /*repeats*/ true, time);
-            }
-            if ((bool)wedBtn.IsChecked)
-            {
-                time = parseRepeatBoxes(wedHrs, wedMins, wedAmPm);
-                alarm.SetRepeat(DayOfWeek.Wednesday, /*repeats*/ true, time);
-            }
-            if ((bool)thuBtn.IsChecked)
-            {
-                time = parseRepeatBoxes(thuHrs, thuMins, thuAmPm);
-                alarm.SetRepeat(DayOfWeek.Thursday, /*repeats*/ true, time);
-            }
-            if ((bool)friBtn.IsChecked)
-            {
-                time = parseRepeatBoxes(friHrs, friMins, friAmPm);
-                alarm.SetRepeat(DayOfWeek.Friday, /*repeats*/ true, time);
-            }
-            if ((bool)satBtn.IsChecked)
-            {
-                time = parseRepeatBoxes(satHrs, satMins, satAmPm);
-                alarm.SetRepeat(DayOfWeek.Saturday, /*repeats*/ true, time);
-            }
         }
     }
 }
