+++ Team4Clock/Team4Clock.csproj
@@ -91,8 +91,13 @@
     </ApplicationDefinition>
     <Compile Include="Alarm.cs" />
     <Compile Include="AlarmRepeats.cs" />
+    <Compile Include="BasicAlarm.cs" />
     <Compile Include="ListOfAlarms.xaml.cs">
       <DependentUpon>ListOfAlarms.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="RepeatingAlarm.cs" />
+    <Compile Include="SetRepeatAlarm.xaml.cs">
+      <DependentUpon>SetRepeatAlarm.xaml</DependentUpon>
     </Compile>
     <Compile Include="SWClock.cs" />
     <Page Include="ListOfAlarms.xaml">
@@ -122,6 +127,10 @@
       <SubType>Code</SubType>
     </Compile>
     <Page Include="SetAlarm.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="SetRepeatAlarm.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
